confirmation box

+++ src/components/layouts/Navbar/navbar.tsx
@@ -6,6 +6,7 @@ import { InputText } from "primereact/inputtext";
 import { MultiSelect } from "primereact/multiselect";
 import { OverlayPanel } from "primereact/overlaypanel";
 import { useNavigate } from "react-router-dom";
+import { ConfirmDialog, confirmDialog } from 'primereact/confirmdialog';
 
 const NavBar = () => {
   const [sidebar, setSidebar] = useState(false);
@@ -27,9 +28,22 @@ const NavBar = () => {
   ];
   return (
     <div className="navbar ">
-      <span></span>
+      <span className="mt-3 cursor-pointer"><i className="pi pi-arrow-left" onClick={(e: any) => {
+            confirmDialog({
+              message: 'Do you want Save Change?',
+              header: 'Confirmation',
+              icon: 'pi pi-info-circle',
+              
+              accept : () =>{
+                localStorage.clear();
+                navigate("/selection");
+              }
+             
+          });
+            
+          }} ></i></span>
       <div
-        className="untitle font-normal"
+        className="untitle font-normal ml-2"
         onClick={(e) => op.current.toggle(e)}
       >
         Untitled <br />
@@ -43,11 +57,22 @@ const NavBar = () => {
         ></i>
       </div>
       <span className="buttonGap">
+      <ConfirmDialog />
         <button
           className="buttonStyle font-normal"
           onClick={(e: any) => {
-            localStorage.clear();
-            navigate("/");
+            confirmDialog({
+              message: 'Do you want to Log Out?',
+              header: 'Log Out Confirmation',
+              icon: 'pi pi-info-circle',
+              
+              accept : () =>{
+                localStorage.clear();
+                navigate("/");
+              }
+             
+          });
+            
           }}
         >
           {" "}
